Avoid referencing traction of previous timestep
This makes the app a lot more convenient to use. It's no longer
necessary to start the simulation at equilibrium, we can apply the load
immediately.

Otherwise this is mathematically equivalent.

The changes in the output of the OGSBenchmark test are due to the fact
that the ramp size in the Neumann conditions (3.333) did not correspond
exactly to the timestep (3.330).

(Precisely the kind of mistake that is now easier to avoid.)