Changed: Compute permeability and dynamic viscosity
... instead of hydraulic conductivity and unit weight.

The proportionality constant in Darcy's law is K/gamma or, equivalently,
kappa/mu. Our input files have been using the former formulation, and the
input called "perm" is confusingly not the permeability.

This switches to the latter form, so that the two inputs are now called
"perm" for permeability, and "mu" for viscosity.